requirements.txt
numpy==1.13.3
tensorflow==1.10.0
Keras==2.2.0
h5py==2.10.0
opencv-python==4.4.0.46
Pillow==8.0.1
matplotlib==3.1.0
scikit-learn==0.24.2

